requirements.txt
# Core dependencies
numpy>=1.21.0
requests>=2.25.0
aiohttp>=3.8.0

# Configuration
hydra-core
omegaconf

# AWS Bedrock
boto3>=1.28.0
botocore>=1.31.0

# Optional OpenRouter support
openai>=1.0.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
